requirements.txt
# core deps
transformers>=4.53.0,<=4.55.0
datasets>=2.16.0,<=3.6.0
accelerate>=1.8.0,<=2.1.0
peft>=0.14.0,<=0.15.2
trl>=0.8.6,<=0.9.6
tokenizers>=0.19.0,<=0.21.1
# advanced training features
torchao>=0.8.0
# gui
gradio>=4.38.0,<=5.42.0
matplotlib>=3.7.0
tyro<0.9.0
# ops
einops
numpy<2.0.0
pandas>=2.0.0
scipy
# model and tokenizer
sentencepiece
tiktoken
modelscope>=1.14.0
hf-transfer
safetensors<=0.5.3
# python
fire
omegaconf
packaging
protobuf
pyyaml
pydantic<=2.10.6
supabase
python-dotenv
# api
uvicorn
fastapi
sse-starlette
# media
av
librosa
